AppComponent: Change behaviour when enabling app permission
When an app permission is disabled from the list, the app permission is then disabled and
it cannot be toggled again in the android app setting.
When it is allowed to be toggled again, it is just meant that the permission is able to be toggled.
It doesn't usually mean that it should be enabled by default.

The change is to make the app permission stays disabled, but it is toggleable.