Introduce a new way to detect new/delete apps
In the old way, the app database is always recreated when launching the app.
Now, if there is no new/deleted apps, the database is untouched. Only app's info will be cached (like before).
The detection is triggered by clicking a HOME tab. So, there is no need to kill the app like before to see the new apps.

minSdkVersion is set to 24 to enable of using Java's stream functionality.

Change-Id: I98c12332d2c33fd98c189585671b791037f1544f
Reviewed-on: http://pi4.lan:8080/c/adhell3/+/3
Reviewed-by: fusionjack <dogfight60-fusionjack@yahoo.de>