Improve new/deleted apps detection
In the previous commit, when there are new/deleted apps, the database is cleared and then the installed apps are inserted into database and the modified apps are restored.
Now, only the new/deleted apps are inserted/deleted to/from database.
For deleted apps: The reported blocked domains and the app's permissions, services and receivers are also deleted from database.

Change-Id: I281a5f2908f6500a886a59be7f2ed4c593f0e69c
Reviewed-on: http://pi4.lan:8080/c/adhell3/+/4
Reviewed-by: fusionjack <dogfight60-fusionjack@yahoo.de>