AppComponent: Check consistency of disabled permissions, services and receivers
When listing app's permissions, services and permissions, check if they are existed in the database.
It is possible that they are disabled on android system, but they are not in the database.
This could occur when an app is reinstalled as we cannot reenable app's components when an app is uninstalled.

Change-Id: I2d562c9859251a1f418924633ac610ea0038760d
Reviewed-on: http://pi4.lan:8080/c/adhell3/+/6
Reviewed-by: fusionjack <dogfight60-fusionjack@yahoo.de>