AppFragment: Using LiveData
LiveData helps to reduce memory leaks and manual lifecycle handling.
It also ensures that UI is always up to date with the data.
The app list is only updated if there is changes in the underlying data, in this case the database.

The swipe-down-refresh is not needed anymore as the list is always up to date with the data.

Change-Id: I72838b01bd730dcd43cbef3246c5e9b39b845a7d
Reviewed-on: http://pi4.lan:8080/c/adhell3/+/16
Reviewed-by: fusionjack <dogfight60-fusionjack@yahoo.de>